backend/requirements.txt
fastapi==0.128.0
uvicorn==0.30.0
sqlalchemy==2.0.46
pydantic==2.12.5
pydantic-settings==2.4.0
python-jose[cryptography]==3.5.0
python-multipart==0.0.7
networkx==3.6.1
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.28.0
python-dotenv==1.0.0
bcrypt==5.0.0
